Separate find utilities and other utilities so we don't break tests
